deep_wide_research/requirements.txt
openai>=1.30.0
anthropic>=0.26.1
google-generativeai>=0.6.0
tavily-python>=0.3.5
exa-py>=1.0.0
pydantic>=2.7.1
httpx>=0.28.1
python-dotenv>=1.0.0
mcp>=1.0.0
